+++ game/conf.lua
@@ -2,4 +2,5 @@ function love.conf(t)
     t.title = "Ludum Dare 57"
     t.version = "11.5" 
     t.author = "The TOL"
+	t.console = true
 end